stocks/requirements.txt
mysql_connector_python==8.3.0
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2
SQLAlchemy==2.0.25
yfinance==0.2.35
python_dotenv==1.0.0
pymongo==4.6.1
